feat: add try-catch to filtering, fix logging, fix linter errors
